+++ src/mandatory/collision_check_dir.c
@@ -1,0 +1,49 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   collision_check.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: minsepar <minsepar@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/04/08 22:51:37 by minsepar          #+#    #+#             */
+/*   Updated: 2024/04/13 16:35:57 by minsepar         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+void	dir_y_check_p(t_map *map, t_user *user, float new_displacement_y)
+{
+	if (map->map[(int)(user->y + new_displacement_y + HITBOX)]
+		[(int)(user->x - HITBOX)] == '0' &&
+	map->map[(int)(user->y + new_displacement_y + HITBOX)]
+		[(int)(user->x + HITBOX)] == '0')
+	user->y += new_displacement_y;
+}
+
+void	dir_y_check_n(t_map *map, t_user *user, float new_displacement_y)
+{
+	if (map->map[(int)(user->y + new_displacement_y - HITBOX)]
+		[(int)(user->x - HITBOX)] == '0' &&
+	map->map[(int)(user->y + new_displacement_y - HITBOX)]
+		[(int)(user->x + HITBOX)] == '0')
+	user->y += new_displacement_y;
+}
+
+void	dir_x_check_p(t_map *map, t_user *user, float new_displacement_x)
+{
+	if (map->map[(int)(user->y - HITBOX)]
+		[(int)(user->x + new_displacement_x + HITBOX)] == '0' &&
+	map->map[(int)(user->y + HITBOX)]
+		[(int)(user->x + new_displacement_x + HITBOX)] == '0')
+	user->x += new_displacement_x;
+}
+
+void	dir_x_check_n(t_map *map, t_user *user, float new_displacement_x)
+{
+	if (map->map[(int)(user->y - HITBOX)]
+		[(int)(user->x + new_displacement_x - HITBOX)] == '0' &&
+	map->map[(int)(user->y + HITBOX)]
+		[(int)(user->x + new_displacement_x - HITBOX)] == '0')
+	user->x += new_displacement_x;
+}
