--- src/mandatory/draw_sprite.c
@@ -1,39 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   draw_sprite.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: minsepar <minsepar@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/12 13:11:40 by jacob             #+#    #+#             */
-/*   Updated: 2024/04/12 19:46:40 by minsepar         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "cub3d.h"
-
-// static void	sort_sprite(t_dda *dda, t_user *user, t_pic *sprite)
-// {
-// 	int	i;
-
-// 	i = -1;
-// 	while (++i < )
-// 	{
-// 		dda->sprite_order[i].index = i;
-// 		dda->sprite_order[i].num
-// 			= (user->x - sprite[i].w) * (user->x - sprite[i].w)
-// 			+ (user->y - sprite[i].h) * (user->y - sprite[i].h);
-// 	}
-// 	quick_sort_sprite(dda->sprite_order, 0,  - 1);
-// }
-
-// void	project_sprite(t_dda *dda, t_user *user, t_pic *sprite)
-// {
-	
-// }
-
-// void	draw_sprite(t_dda *dda, t_mlx *graphic, t_user *user)
-// {
-// 	sort_sprite(dda, user, graphic->sprite);
-// 	project_sprite(dda, user, graphic->sprite);
-// }
