+++ src/mandatory/frame.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   frame.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jacob <jacob@student.42.fr>                +#+  +:+       +#+        */
+/*   By: minsepar <minsepar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/03 22:57:58 by minsepar          #+#    #+#             */
-/*   Updated: 2024/04/11 22:47:12 by jacob            ###   ########.fr       */
+/*   Updated: 2024/04/13 17:50:40 by minsepar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,10 +24,7 @@ void	display_frame(t_mlx *graphic)
 {
 	t_data	*data;
 
-	// printf("frame_num: [%d]\n", graphic->num_frame);
 	data = &graphic->img_data[graphic->num_frame];
-	// printf("data->img [%p]\n", data->img);
 	mlx_put_image_to_window(graphic->mlx, graphic->win, data->img, 0, 0);
-	// printf("x: [%f] y: [%f]\n", graphic->user.x, graphic->user.y);
 	graphic->num_frame ^= 1;
 }
