+++ src/mandatory/mlx_hooks.c
@@ -6,7 +6,7 @@
 /*   By: minsepar <minsepar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/03 15:14:28 by minsepar          #+#    #+#             */
-/*   Updated: 2024/04/13 16:58:56 by minsepar         ###   ########.fr       */
+/*   Updated: 2024/04/13 17:46:52 by minsepar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
