+++ src/mandatory/mlx_pixel.c
@@ -6,7 +6,7 @@
 /*   By: minsepar <minsepar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/03 14:45:09 by minsepar          #+#    #+#             */
-/*   Updated: 2024/04/13 16:59:47 by minsepar         ###   ########.fr       */
+/*   Updated: 2024/04/13 17:52:01 by minsepar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
